+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Unreleased
 **Bugfixes**
  - Fixed DS2 score code for CAS that was generated when registering a Python model.
  - `PyMAS.score_code(dest='ESP')` corrected to `dest='EP'`
+ - Fixed an issue where long user-defined properties prevented model registration.
  
  
 v1.1.1 (2019-8-6)